worker/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1
python-dotenv==1.0.1
pypdf==4.0.1
tiktoken==0.5.2
openai==1.14.3